# docs/swagger.yaml
basePath: /api/v1
definitions:
  features_statistics_transport_http.GetStatisticsResponse:
    properties:
      task_average_completion_time:
        type: string
      task_completed:
        type: integer
      task_completed_rate:
        type: number
      tasks_created:
        type: integer
    type: object
  features_tasks_transport_http.CreateTaskRequest:
    properties:
      author_user_id:
        type: integer
      description:
        maxLength: 1000
        minLength: 1
        type: string
      title:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - author_user_id
    - title
    type: object
  features_tasks_transport_http.CreateTaskResponse:
    properties:
      author_user_id:
        type: integer
      completed:
        type: boolean
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  features_tasks_transport_http.GetTaskResponse:
    properties:
      author_user_id:
        type: integer
      completed:
        type: boolean
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  features_tasks_transport_http.PatchTaskRequest:
    properties:
      completed:
        $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_types.Nullable-bool'
      description:
        $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_types.Nullable-string'
      title:
        $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_types.Nullable-string'
    type: object
  features_tasks_transport_http.PatchUserResponse:
    properties:
      author_user_id:
        type: integer
      completed:
        type: boolean
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  features_tasks_transport_http.TaskDTOResponse:
    properties:
      author_user_id:
        type: integer
      completed:
        type: boolean
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  features_users_transport_http.CreateUserRequest:
    properties:
      full_name:
        example: Ivan Ivanov
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        example: "+79997776767"
        type: string
    required:
    - full_name
    type: object
  features_users_transport_http.CreateUserResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79997776767"
        type: string
      version:
        example: 3
        type: integer
    type: object
  features_users_transport_http.GetUserResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79997776767"
        type: string
      version:
        example: 3
        type: integer
    type: object
  features_users_transport_http.PatchUserRequest:
    properties:
      full_name:
        $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_types.Nullable-string'
      phone_number:
        $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_types.Nullable-string'
    type: object
  features_users_transport_http.PatchUserResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79997776767"
        type: string
      version:
        example: 3
        type: integer
    type: object
  features_users_transport_http.UserDTOResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79997776767"
        type: string
      version:
        example: 3
        type: integer
    type: object
  github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short human-readable message
        type: string
    type: object
  github_com_raizonw_todo-app_internal_core_transport_http_types.Nullable-bool:
    properties:
      set:
        type: boolean
      value:
        type: boolean
    type: object
  github_com_raizonw_todo-app_internal_core_transport_http_types.Nullable-string:
    properties:
      set:
        type: boolean
      value:
        type: string
    type: object
host: 127.0.0.1:5050
info:
  contact: {}
  description: Todo Application REST-API scheme
  title: Golang Todo API
  version: "1.0"
paths:
  /statistics:
    get:
      consumes:
      - application/json
      description: Получить статистику по задачам пользователя за определенный период
        (создано, выполнено, среднее время выполнения)
      parameters:
      - description: ID пользователя
        in: query
        name: user_id
        type: integer
      - description: Дата начала периода (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: Дата окончания периода (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Статистика задач
          schema:
            $ref: '#/definitions/features_statistics_transport_http.GetStatisticsResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Получить статистику задач
      tags:
      - statistics
  /tasks:
    get:
      consumes:
      - application/json
      description: Получить список задач с фильтрацией по user_id и поддержкой пагинации
        (limit/offset)
      parameters:
      - description: ID пользователя-автора задач
        in: query
        name: user_id
        type: integer
      - description: Лимит количества задач
        in: query
        name: limit
        type: integer
      - description: Смещение относительно начала
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список задач
          schema:
            items:
              $ref: '#/definitions/features_tasks_transport_http.TaskDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Получить список задач
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Создать новую задачу для пользователя
      parameters:
      - description: Тело запроса на создание задачи
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/features_tasks_transport_http.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданная задача
          schema:
            $ref: '#/definitions/features_tasks_transport_http.CreateTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Создать задачу
      tags:
      - tasks
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Удалить задачу из системы по её ID
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Успешное удаление (нет содержимого)
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Удалить задачу
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: Получить подробную информацию о задаче по её ID
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Информация о задаче
          schema:
            $ref: '#/definitions/features_tasks_transport_http.GetTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Получить задачу
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: Частично обновить информацию о задаче (название, описание, статус
        завершения)
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      - description: Тело запроса на обновление
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/features_tasks_transport_http.PatchTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная информация о задаче
          schema:
            $ref: '#/definitions/features_tasks_transport_http.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Обновить задачу
      tags:
      - tasks
  /users:
    get:
      consumes:
      - application/json
      description: Получить список пользователей с поддержкой пагинации (limit/offset)
      parameters:
      - description: Лимит количества пользователей
        in: query
        name: limit
        type: integer
      - description: Смещение относительно начала
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список пользователей
          schema:
            items:
              $ref: '#/definitions/features_users_transport_http.UserDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Получить список пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Создать нового пользователя в системе
      parameters:
      - description: CreateUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/features_users_transport_http.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданный пользователь
          schema:
            $ref: '#/definitions/features_users_transport_http.CreateUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Создать пользователя
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Удалить пользователя из системы по его ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Успешное удаление (нет содержимого)
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Удалить пользователя
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Получить информацию о пользователе по его ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Информация о пользователе
          schema:
            $ref: '#/definitions/features_users_transport_http.GetUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Получить пользователя
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Частично обновить информацию о пользователе (имя и/или номер телефона)
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Тело запроса на обновление
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/features_users_transport_http.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная информация о пользователе
          schema:
            $ref: '#/definitions/features_users_transport_http.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_raizonw_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Обновить пользователя
      tags:
      - users
swagger: "2.0"